Tratamento de exceções do delete e do update

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,17 +1,18 @@
 package com.devsuperior.dscommerce.services;
 
-import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDto;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -50,19 +51,29 @@ public class ProductService {
 	
 	@Transactional 
 	public ProductDto update(Long id, ProductDto dto) {
-		
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto,entity);
-		entity = repository.save(entity);
-		return new ProductDto(entity);
+		try {
+			
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto,entity);
+			entity = repository.save(entity);
+			return new ProductDto(entity);
+		}catch (Exception e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		
-		repository.deleteById(id);
-		
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
